+++ bin/faucet/src/faucet/mod.rs
@@ -25,7 +25,7 @@ use miden_objects::{
     },
     vm::AdviceMap,
 };
-use miden_proving_service_client::proving_service::tx_prover::RemoteTransactionProver;
+use miden_remote_prover_client::remote_prover::tx_prover::RemoteTransactionProver;
 use miden_tx::{
     LocalTransactionProver, ProvingOptions, TransactionExecutor, TransactionExecutorError,
     TransactionProver, TransactionProverError, auth::BasicAuthenticator,