+++ crates/block-producer/src/errors.rs
@@ -7,7 +7,7 @@ use miden_objects::{
     note::{NoteId, Nullifier},
     transaction::TransactionId,
 };
-use miden_proving_service_client::RemoteProverError;
+use miden_remote_prover_client::RemoteProverClientError;
 use thiserror::Error;
 use tokio::task::JoinError;
 
@@ -146,7 +146,7 @@ pub enum BuildBatchError {
     ProveBatchError(#[source] ProvenBatchError),
 
     #[error("failed to prove batch with remote prover")]
-    RemoteProverError(#[source] RemoteProverError),
+    RemoteProverClientError(#[source] RemoteProverClientError),
 
     #[error("batch proof security level is too low: {0} < {1}")]
     SecurityLevelTooLow(u32, u32),
@@ -170,7 +170,7 @@ pub enum BuildBlockError {
     #[error("nothing actually went wrong, failure was injected on purpose")]
     InjectedFailure,
     #[error("failed to prove block with remote prover")]
-    RemoteProverError(#[source] RemoteProverError),
+    RemoteProverClientError(#[source] RemoteProverClientError),
     #[error("block proof security level is too low: {0} < {1}")]
     SecurityLevelTooLow(u32, u32),
 }