+++ crates/ntx-builder/src/builder/prover.rs
@@ -1,5 +1,5 @@
 use miden_objects::transaction::{ExecutedTransaction, TransactionWitness};
-use miden_proving_service_client::proving_service::tx_prover::RemoteTransactionProver;
+use miden_remote_prover_client::remote_prover::tx_prover::RemoteTransactionProver;
 use miden_tx::{LocalTransactionProver, TransactionProver};
 use url::Url;