+++ CHANGELOG.md
@@ -27,6 +27,10 @@
 - Faucet challenge expiration time is now configurable (#1017).
 - Removed system monitor from node binary (#1019).
 - [BREAKING] Renamed `open_telemetry` to `enable_otel` in all node's commands (#1019).
+- [BREAKING] Rename `miden-proving-service` to `miden-remote-prover` (#1004).
+- [BREAKING] Rename `miden-proving-service-client` to `miden-remote-prover-client` (#1004).
+- [BREAKING] Rename `RemoteProverError` to `RemoteProverClientError` (#1004).
+- [BREAKING] Rename `ProvingServiceError` to `RemoteProverError` (#1004).
 
 ### Fixes